raycast

+++ headers/ft_transform.h
@@ -6,7 +6,7 @@
 /*   By: ezequeil <ezequeil@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/19 11:12:48 by edos-san          #+#    #+#             */
-/*   Updated: 2022/07/03 20:03:08 by ezequeil         ###   ########.fr       */
+/*   Updated: 2022/07/04 22:17:52 by ezequeil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,14 +18,14 @@ typedef struct s_vector		t_vector;
 
 struct s_vector
 {
-	int			x;
-	int			y;
-	int			z;
-	int			w;
-	int			h;
+	float			x;
+	float			y;
+	float			z;
+	float			w;
+	float			h;
 };
 
-t_vector	vector(int x, int y, int w, int h);
+t_vector	vector(float x, float y, float w, float h);
 t_vector	vector_zero(void);
 
 #endif
