+++ src/objects/map/ft_map_util.c
@@ -6,7 +6,7 @@
 /*   By: ezequeil <ezequeil@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/20 22:00:43 by edos-san          #+#    #+#             */
-/*   Updated: 2022/07/01 23:20:00 by ezequeil         ###   ########.fr       */
+/*   Updated: 2022/07/04 22:19:07 by ezequeil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,11 +44,11 @@ void	print_check_map(t_map *m)
 			else if (m->dir == 1)
 				dir = v.y / 32;
 			else if (m->dir == 2)
-				dir = m->check[v.y / 32][v.x / 32] == 1;
+				dir = m->check[(int)v.y / 32][(int)v.x / 32] == 1;
 			b[0] = ('0' + dir / 10);
 			b[1] = ('0' + (dir % 10));
-			if ((m->dir != 2 || dir) && m->maps[v.y / 32][v.x / 32] \
-			!= '1' && m->maps[v.y / 32][v.x / 32] != ' ')
+			if ((m->dir != 2 || dir) && m->maps[(int)v.y / 32][(int)v.x / 32] \
+			!= '1' && m->maps[(int)v.y / 32][(int)v.x / 32] != ' ')
 				mlx_string_put(engine()->mlx, engine()->win, v.x + 5, v.y + 5, \
 				0x111111, b);
 			v.x += 32;
